refactor(hew-lsp): extract language feature handlers
