+++ hew-lsp/src/server/handlers/mod.rs
@@ -1,4 +1,5 @@
 pub(super) mod hierarchy;
+pub(super) mod language_features;
 pub(super) mod navigation;
 pub(super) mod text_sync;
 pub(super) mod workspace;
